requirements.txt
boto3>=1.23.5
botocore==1.26.5
discord==1.7.3
discord.py==1.7.3
python-dateutil>=2.8.2
